auth-service/requirements.txt
# pydantic
pydantic[email]==2.5.2
pydantic-settings==2.1.0

# asynchronous api
fastapi==0.104.1
orjson==3.9.10
uvicorn==0.12.2
gunicorn==21.2.0
#uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"

# databases
psycopg==3.1.17
psycopg-binary==3.1.17
psycopg2
redis==5.0.1
backoff==2.2.1
alembic==1.13.1
SQLAlchemy==2.0.25
asyncpg==0.29.0
async-sqlalchemy==1.0.0

#passwords
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6

#backoff
backoff==2.2.1

# yandex
itsdangerous==2.1.2
httpx==0.26.0
authlib==0.14.3
